Provide per section attachments
Upload and delete attachments on a per section basis.
