+++ gui/app/styles/core/view/document/section-editor.scss
@@ -4,10 +4,6 @@
 		box-shadow: 0 0 0 0.75pt map-get($gray-shades, 200),0 0 3pt 0.75pt map-get($gray-shades, 200);
 		border: 1px solid map-get($gray-shades, 200);
 	}
-
-	> .attachments {
-		
-	}
 }
 
 .content-linker-modal-container {
