+++ gui/app/templates/components/document/document-editor.hbs
@@ -1,1 +1,7 @@
-{{component editorType document=document folder=folder page=page meta=meta onCancel=(action "onCancel") onAction=(action "onAction")}}
+{{component editorType 
+    document=document 
+    folder=folder 
+    page=page 
+    meta=meta 
+    attachments=attachments
+    onCancel=(action "onCancel") onAction=(action "onAction")}}
