+++ gui/app/templates/components/document/view-content.hbs
@@ -12,13 +12,16 @@
 			document=document
 			pending=item.pending
 			permissions=permissions
+			attachments=attachments
 			refresh=(action refresh)
+			onAttachmentUpload=(action onAttachmentUpload)
+			onAttachmentDelete=(action onAttachmentDelete)
 			onSavePage=(action "onSavePage")
 			onCopyPage=(action "onCopyPage")
 			onMovePage=(action "onMovePage")
 			onDeletePage=(action "onDeletePage")
 			onSavePageAsBlock=(action "onSavePageAsBlock")
-			onPageLevelChange=(action onPageLevelChange)
+			onPageLevelChange=(action onPageLevelChange)			
 			onPageSequenceChange=(action onPageSequenceChange)
 			onShowSectionWizard=(action "onShowSectionWizard")}}
 	{{/each}}
