+++ gui/app/templates/components/section/base-editor-inline.hbs
@@ -41,12 +41,6 @@
 		<div class="canvas">
 			{{yield}}
 		</div>
-
-		<div class="attachments">
-			{{ui/ui-spacer size=100}}
-			{{ui/ui-button color=constants.Color.Gray label="Upload" id=uploadId}}
-			{{ui/ui-spacer size=100}}
-		</div>
 	</div>
 </div>
